requirements.txt
pillow
pytesseract
opencv-python
numpy
#easyocr
tabulate[widechars]
requests
